requirements.txt
streamlit==1.35.0
openai==1.27.0
langchain==0.1.14
python-dotenv==1.0.1
